requirements.txt
attrs==20.3.0
click==7.1.2
Flask==1.1.2
Flask-HTTPAuth==4.2.0
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
kube-manifest-lint==0.2.0
ldap3==2.9
MarkupSafe==1.1.1
pyasn1==0.4.8
pyrsistent==0.17.3
PyYAML==5.4.1
six==1.15.0
Werkzeug==1.0.1
